* commit 'd0788b5c3a5eb3b7b71b2c68715a1da9562d0e35':
  ipsec: fix secondary partition id for ls1088a
  ipsec: klocwork uint cast fixes

+++ src/drivers/netf/ipsec/00_01/ipsec.c
@@ -3293,7 +3293,7 @@ IPSEC_CODE_PLACEMENT uint8_t ipsec_get_ipv6_nh_offset(
 
 	/* IP Header Length for SEC encapsulation, including IP header and
 	 * extensions before ESP */
-	*length = sizeof(struct ipv6hdr);
+	*length = (uint8_t)sizeof(struct ipv6hdr);
 
 	/* From RFC 2460: 4.1  Extension Header Order
 	 * IPv6 nodes must accept and attempt to process extension headers in
@@ -3678,7 +3678,7 @@ int ipsec_force_seconds_lifetime_expiry(
 			cdma_read(
 				&sap2, /* void *ws_dst */
 				IPSEC_SA_PARAMS_2_ADDR(desc_addr),
-				sizeof(sap2) /* uint16_t size */
+				(uint16_t)sizeof(sap2) /* uint16_t size */
 			);
 		}
